Move file into the remote package
